+++ orthanc_tools/__init__.py
@@ -8,6 +8,8 @@ from .orthanc_comparator import OrthancComparator
 from .orthanc_cleaner import OrthancCleaner
 from .orthanc_replicator import OrthancReplicator
 from .ids_migrator import IdsMigrator
+from .orthanc_downloader import *
+from .orthanc_warmer import OrthancWarmer
 
 from .hl7Lib import *
 from .helpers import *